# backend/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ASTRO BACKEND API",
    "version": "1.0.0",
    "description": "API documentation using Swagger"
  },
  "servers": [
    {
      "url": "http://ec2-3-106-223-174.ap-southeast-2.compute.amazonaws.com"
    },
    {
      "url": "https://meetastro.onrender.com"
    },
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
  "paths": {
...
    "/api/query/generate-voice-answer": {
      "post": {
        "summary": "Generate a voice response based on user query",
        "operationId": "generateVoiceAnswer",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "query": {
                    "type": "string",
                    "description": "The user's query for which a voice response is generated."
                  },
                  "ip": {
                    "type": "string",
                    "description": "The IP address of the user."
                  }
                },
                "required": ["query", "ip"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response with the generated voice answer",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "botResponse": {
                      "type": "string",
                      "description": "The generated voice response."
                    },
                    "audioUrl": {
                      "type": "string",
                      "description": "URL to the audio file of the response."
                    },
                    "audioDuration": {
                      "type": "integer",
                      "description": "Duration of the audio in seconds."
                    },
                    "remaining": {
                      "type": "integer",
                      "description": "Number of remaining requests for the user."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message describing the issue."
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests - Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "botResponse": {
                      "type": "string",
                      "description": "Message indicating the rate limit has been reached."
                    },
                    "remaining": {
                      "type": "integer",
                      "description": "Number of remaining requests (0 in this case)."
                    },
                    "audioUrl": {
                      "type": "string",
                      "description": "URL to the audio file explaining the rate limit."
                    },
                    "audioDuration": {
                      "type": "integer",
                      "description": "Duration of the audio in seconds."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An error occurred while processing the request.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message indicating a server error."
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}